LPS-148315 Moving useInitialPanelState hook closer to usage
