--- modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/constants.js
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- */
-
-export const ANALYTICS_REPORTS_CLOSED_PANEL_VALUE = 'closed';
-export const ANALYTICS_REPORTS_OPEN_PANEL_VALUE = 'open';
-export const ANALYTICS_REPORTS_PANEL_ID =
-	'com.liferay.analytics.reports.web_panelState';
